--- a/modules/dicom_archive/php/dicomarchiveanonymizer.class.inc
+++ b/modules/dicom_archive/php/dicomarchiveanonymizer.class.inc
@@ -48,7 +48,7 @@ class DICOMArchiveAnonymizer implements Mapper
             $config = \NDB_Config::singleton()->getSetting("imaging_modules");
         }
 
-        $newrow = $resource->ToArray();
+        $newrow = json_decode($resource->toJSON(), true);
         $cid    = $resource->getCenterID();
 
         if (!preg_match($config['patientNameRegex'], $newrow['Patient Name'])