--- a/modules/dicom_archive/php/dicomarchiverowprovisioner.class.inc
+++ b/modules/dicom_archive/php/dicomarchiverowprovisioner.class.inc
@@ -45,6 +45,8 @@ class DicomArchiveRowProvisioner extends \LORIS\Data\Provisioners\DBRowProvision
             t.ArchiveLocation as Archive_Location,
             'View Details' as Metadata,
             'View Images' as `MRI Browser`,
+            (SELECT GROUP_CONCAT(SeriesUID) FROM tarchive_series e WHERE
+                    e.TarchiveID=t.TarchiveID GROUP BY e.TarchiveID) as SeriesUID,
             s.CenterID as Site,
             t.TarchiveID as TarchiveID,
             s.ID as SessionID,