--- a/src/Data/Mapper.php
+++ b/src/Data/Mapper.php
@@ -44,5 +44,5 @@ interface Mapper
      *
      * @return DataInstance a new DataInstance with the map applied.
      */
-    function map(\User $user, DataInstance $resource) : Instance;
+    function map(\User $user, DataInstance $resource) : DataInstance;
 }