This updates the DICOM Archive module to use the new DICOM Archive framework as a first use case. It removes the custom generated NDB_Menu_Filter SQL and converts it to a DBRowProvisioner.

Note that I'm not entirely happy with the amount of serializing and deserializing from JSON required to modify/annotate responses at various parts (because of the fact that the interface only has toJSON and no array equivalent, while the pre-interface class had both toArray and toJSON, so it could delay the serialization until the last minute), but this still appears to perform better than the NDB_Menu_Filter variation (with close to the same query) under all circumstances because of the usage of lazily evaluated PDOStatements in DBRowProvisioner vs eagerly evaluated arrays of `Database->pselect` deep inside the NDB_Menu_Filter.

With the first benchmarking tool I could find (I used https://github.com/wg/wrk), the following results occurred (note that these are using the IBIS dataset on my 3 year old laptop with the laptop running MySQL, a VM with Apache/LORIS mounting the LORIS directory from the host, and the benchmark all locally on the laptop. It's a realistic/large data set, but not a realistic production environment. Numbers should be compared for relative performance only...):

aces/major without this change (light load,  1 concurrent connection):

```
Running 30s test @ http://192.168.57.101/dicom_archive/?format=json
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   286.76ms   32.83ms 609.46ms   97.14%
    Req/Sec     3.19      0.46     4.00     77.88%
  Latency Distribution
     50%  281.17ms
     75%  284.80ms
     90%  295.63ms
     99%  320.97ms
```

With the change (light load, 1 concurrent connection)
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   176.82ms   11.78ms 266.93ms   89.29%
    Req/Sec     5.58      2.59    10.00     72.62%
  Latency Distribution
     50%  174.17ms
     75%  177.48ms
     90%  188.17ms
     99%  234.26ms
  168 requests in 30.04s, 234.05MB read
```

With an unrealistic (for the dicom_archive) load of 100 concurrent connections , the results are embarrassing under both the status quo and the new framework, but show a definite improvement with the modifications:

"High" load (100 concurrent connections):

Before changes:
```
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.57s   210.08ms   1.71s   100.00%
    Req/Sec     2.29      1.28     5.00     64.13%
  Latency Distribution
     50%    1.71s 
     75%    1.71s 
     90%    1.71s 
     99%    1.71s 
  92 requests in 30.02s, 207.80MB read
  Socket errors: connect 0, read 0, write 0, timeout 90
```

After changes:
```
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.33s   446.80ms   1.97s    62.50%
    Req/Sec     2.53      1.89    10.00     83.09%
  Latency Distribution
     50%    1.41s 
     75%    1.80s 
     90%    1.97s 
     99%    1.97s 
  136 requests in 30.10s, 189.47MB read
  Socket errors: connect 0, read 0, write 0, timeout 128
```

In general, timeouts start dominating the statistics after an embarrassingly low 7 concurrent connections with the current NDB_Menu_Filter model, and a slightly better but still embarrassingly low 12 after the changes.

The statistics at 7 (which is probably still an unrealistically high load for the dicom archive in the average LORIS instance):

Moderate/"highest" load:

Before change:
```
Running 30s test @ http://192.168.57.101/dicom_archive/?format=json
  4 threads and 7 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.11s    84.27ms   1.25s    96.23%
    Req/Sec     0.03      0.22     2.00     98.11%
  Latency Distribution
     50%    1.12s 
     75%    1.13s 
     90%    1.14s 
     99%    1.16s 
  106 requests in 30.07s, 239.59MB read
```

After change:
```
Running 30s test @ http://192.168.57.101/dicom_archive/?format=json
  4 threads and 7 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   692.05ms   45.06ms 831.91ms   95.35%
    Req/Sec     1.01      0.15     3.00     99.42%
  Latency Distribution
     50%  692.52ms
     75%  700.11ms
     90%  708.21ms
     99%  827.24ms
  172 requests in 30.11s, 239.62MB read
```