@driusan 2018-03-01
@kongtiaowang I manually tested this and there's not exactly a lot of functionality to test, but if you could give it a manual test too before merging that would be useful..