@kongtiaowang 2018-03-01
@driusan  Can't load dicom_archive page.
![screen shot 2018-03-01 at 12 34 08 pm](https://user-images.githubusercontent.com/9876007/36859605-e38b9e76-1d4c-11e8-8526-1d1f35b967b0.png)