@PapillonMcGill 2018-03-05
Code look clean, haven't spotted anything by looking at it.